feat: add dialog for activating online play
