+++ src/renderer/containers/Header/ActivateOnlineDialog.tsx
@@ -8,13 +8,15 @@ import { useTheme } from "@material-ui/core/styles";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
 import React from "react";
 
-export interface StartGameDialogProps {
+import { ActivateOnlineForm } from "../ActivateOnlineForm";
+
+export interface ActivateOnlineDialogProps {
   open: boolean;
   onClose: () => void;
   onSubmit: () => void;
 }
 
-export const StartGameDialog: React.FC<StartGameDialogProps> = ({ open, onClose, onSubmit }) => {
+export const ActivateOnlineDialog: React.FC<ActivateOnlineDialogProps> = ({ open, onClose, onSubmit }) => {
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down("xs"));
 
@@ -27,16 +29,16 @@ export const StartGameDialog: React.FC<StartGameDialogProps> = ({ open, onClose,
   return (
     <Dialog open={open} onClose={onClose} fullWidth={true} fullScreen={fullScreen} disableBackdropClick={true}>
       <form onSubmit={handleSubmit}>
-        <StyledDialogTitle>You are not logged in</StyledDialogTitle>
+        <StyledDialogTitle>Online play is disabled</StyledDialogTitle>
         <DialogContent style={{ display: "flex" }}>
-          <div>Only logged in users can play online. Would you like to play offline?</div>
+          <ActivateOnlineForm hideRetry={true} />
         </DialogContent>
         <DialogActions>
           <Button onClick={onClose} color="secondary">
             Cancel
           </Button>
           <Button color="primary" type="submit">
-            Play offline
+            Retry
           </Button>
         </DialogActions>
       </form>
